requirements.txt
playwright
supabase-py
psycopg2-binary
scikit-learn
openai
python-dotenv
streamlit
pandas